Updated serverless test suites to new structure
