+++ tests/suites/serverless/queue/default/stack/Pulumi.yaml
