+++ tests/suites/serverless/queue/default/stack/package.json
