+++ tests/suites/serverless/queue/default/stack/tsconfig.json
